+++ src/CombatDicesTeam.Combats/CombatantEffectLifetimes/IsAttackUsedLifetimeExpirationCondition.cs
@@ -3,9 +3,9 @@ using JetBrains.Annotations;
 namespace CombatDicesTeam.Combats.CombatantEffectLifetimes;
 
 [PublicAPI]
-public sealed class IsAttackCombatMovePredicate : ICombatMovePredicate
+public sealed class IsAttackUsedLifetimeExpirationCondition : IUsedCombatMovementLifetimeExpirationCondition
 {
-    public bool Check(CombatMovementInstance combatMove)
+    public bool Check(ICombatant statusOwner, CombatMovementInstance combatMove)
     {
         return combatMove.SourceMovement.Tags.HasFlag(CombatMovementTags.Attack);
     }