+++ src/CombatDicesTeam.Combats/CombatantStatuses/AlliesAuraTargetSelector.cs
@@ -3,10 +3,10 @@ using JetBrains.Annotations;
 namespace CombatDicesTeam.Combats.CombatantStatuses;
 
 /// <summary>
-/// Aura imposes to all enemies.
+/// Aura imposes to all allies.
 /// </summary>
 [PublicAPI]
-public sealed class EnemiesAuraTargetSelector : IAuraTargetSelector
+public sealed class AlliesAuraTargetSelector : IAuraTargetSelector
 {
     public bool IsCombatantUnderAura(ICombatant auraOwner, ICombatant testCombatant, IAuraTargetSelectorContext context)
     {