+++ src/CombatDicesTeam.Combats/CombatantStatuses/AuraCombatantStatus.cs
@@ -9,7 +9,8 @@ public sealed class AuraCombatantStatus : CombatantStatusBase
     private ICombatant? _owner;
 
     public AuraCombatantStatus(ICombatantStatusSid sid, ICombatantStatusLifetime lifetime,
-        ICombatantStatusSource source, Func<ICombatant, ICombatantStatusFactory> auraStatusDelegate, IAuraTargetSelector auraTargetSelector) :
+        ICombatantStatusSource source, Func<ICombatant, ICombatantStatusFactory> auraStatusDelegate,
+        IAuraTargetSelector auraTargetSelector) :
         base(sid, lifetime, source)
     {
         _auraStatusDelegate = auraStatusDelegate;
@@ -23,8 +24,10 @@ public sealed class AuraCombatantStatus : CombatantStatusBase
         _owner = combatant;
         _combat = context.Combat;
 
-        var auraTargets =
-            context.Combat.CurrentCombatants.Where(x => _auraTargetSelector.IsCombatantUnderAura(x, combatant));
+        var auraContext = new AuraTargetSelectorContext(_combat);
+
+        var auraTargets = context.Combat.CurrentCombatants
+            .Where(x => _auraTargetSelector.IsCombatantUnderAura(_owner, x, auraContext));
 
         // Add status to current combatants
         foreach (var target in auraTargets)
@@ -57,7 +60,9 @@ public sealed class AuraCombatantStatus : CombatantStatusBase
 #endif
             }
 
-            if (!_auraTargetSelector.IsCombatantUnderAura(_owner, combatant))
+            var auraContext = new AuraTargetSelectorContext(_combat);
+
+            if (!_auraTargetSelector.IsCombatantUnderAura(_owner, combatant, auraContext))
             {
                 // Do not add status new allies
                 return;