+++ src/CombatDicesTeam.Combats/CombatantStatuses/IAuraTargetSelector.cs
@@ -2,5 +2,5 @@ namespace CombatDicesTeam.Combats.CombatantStatuses;
 
 public interface IAuraTargetSelector
 {
-    bool IsCombatantUnderAura(ICombatant auraOwner, ICombatant testCombatant);
+    bool IsCombatantUnderAura(ICombatant auraOwner, ICombatant testCombatant, IAuraTargetSelectorContext context);
 }