+++ src/CombatDicesTeam.Combats/Effects/InterruptEffect.cs
@@ -9,7 +9,7 @@ public class InterruptEffect : IEffect
 
     public ITargetSelector Selector { get; }
 
-    public IReadOnlyCollection<IEffectCondition> ImposeConditions => Array.Empty<IEffectCondition>();
+    public IReadOnlyCollection<IEffectCondition> ImposeConditions { get; init; } = Array.Empty<IEffectCondition>();
 
     public IEffectInstance CreateInstance()
     {