+++ src/CombatDicesTeam.Combats/Effects/PeriodicEffect.cs
@@ -14,7 +14,7 @@ public sealed class PeriodicEffect : IEffect
 
     public ITargetSelector Selector { get; }
 
-    public IReadOnlyCollection<IEffectCondition> ImposeConditions => Array.Empty<IEffectCondition>();
+    public IReadOnlyCollection<IEffectCondition> ImposeConditions { get; init; } = Array.Empty<IEffectCondition>();
 
     public IEffectInstance CreateInstance()
     {